02_PPO/requirements.txt
# Python dependencies for PPO baseline

# Core dependencies
stable-baselines3[extra]==2.2.1
gymnasium==0.29.1
torch>=1.13.0
numpy>=1.21.0

# For Isaac Gym integration (if using custom env)
# Note: Isaac Gym should already be installed in your main environment

# Monitoring and logging
tensorboard>=2.11.0
matplotlib>=3.5.0
seaborn>=0.12.0

# Optional: for better performance
# opencv-python>=4.6.0  # for video recording
